+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 Format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+---
+
+## [1.2.0] - 2025-07-27
+
+### Added
+- Added `-v` flag to display version info
+- Added test framework, test cases, test key & test results
+- Added contributing guidelines
+
+### Fixed
+- Removed additional key file output when encrypting / decrypting to stdout because it breaks easy redirecting / piping
+
+### Changed
+- Update documentation
+
+---
 
 ## [1.1.0] - 2025-07-25